refactor(ast): share raw shell scanner
