+++ crates/shuck-formatter/src/scan.rs
@@ -1,4 +1,5 @@
 use shuck_ast::Span;
+pub(crate) use shuck_ast::raw_shell::{QuoteState, RawShellScanner, shell_comment_can_start};
 
 use crate::comments::SourceMap;
 
@@ -273,385 +274,6 @@ fn is_shell_keyword_char(ch: char) -> bool {
     ch.is_ascii_alphanumeric() || ch == '_'
 }
 
-pub(crate) fn shell_comment_can_start(source: &str, offset: usize) -> bool {
-    source[..offset]
-        .chars()
-        .next_back()
-        .is_none_or(|ch| ch == '\n' || ch.is_whitespace() || matches!(ch, ';' | '&' | '|'))
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RawQuoteKind {
-    Single,
-    Double,
-    Backtick,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub(crate) struct QuoteState {
-    quote: Option<RawQuoteKind>,
-    escaped: bool,
-}
-
-impl QuoteState {
-    pub(crate) fn in_multiline_literal(&self) -> bool {
-        self.quote.is_some()
-    }
-
-    pub(crate) fn in_single_quotes(&self) -> bool {
-        self.quote == Some(RawQuoteKind::Single)
-    }
-
-    pub(crate) fn in_quotes(&self) -> bool {
-        self.quote.is_some()
-    }
-
-    pub(crate) fn consume_raw_char(&mut self, ch: char, include_backticks: bool) -> bool {
-        if self.escaped {
-            self.escaped = false;
-            return true;
-        }
-
-        match self.quote {
-            Some(RawQuoteKind::Single) => {
-                if ch == '\'' {
-                    self.quote = None;
-                }
-                true
-            }
-            Some(RawQuoteKind::Double) => {
-                if ch == '"' {
-                    self.quote = None;
-                } else if ch == '\\' {
-                    self.escaped = true;
-                }
-                true
-            }
-            Some(RawQuoteKind::Backtick) if include_backticks => {
-                if ch == '`' {
-                    self.quote = None;
-                } else if ch == '\\' {
-                    self.escaped = true;
-                }
-                true
-            }
-            _ if ch == '\\' => {
-                self.escaped = true;
-                true
-            }
-            _ if ch == '\'' => {
-                self.quote = Some(RawQuoteKind::Single);
-                true
-            }
-            _ if ch == '"' => {
-                self.quote = Some(RawQuoteKind::Double);
-                true
-            }
-            _ if include_backticks && ch == '`' => {
-                self.quote = Some(RawQuoteKind::Backtick);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    pub(crate) fn consume_shell_word_char(&mut self, ch: char) -> bool {
-        if self.escaped {
-            self.escaped = false;
-            return true;
-        }
-
-        match self.quote {
-            Some(RawQuoteKind::Single) => {
-                if ch == '\'' {
-                    self.quote = None;
-                }
-                true
-            }
-            Some(RawQuoteKind::Double) => {
-                if ch == '"' {
-                    self.quote = None;
-                    return true;
-                }
-                if ch == '\\' {
-                    self.escaped = true;
-                    return true;
-                }
-                false
-            }
-            _ if ch == '\\' => {
-                self.escaped = true;
-                true
-            }
-            _ if ch == '\'' => {
-                self.quote = Some(RawQuoteKind::Single);
-                true
-            }
-            _ if ch == '"' => {
-                self.quote = Some(RawQuoteKind::Double);
-                true
-            }
-            _ => false,
-        }
-    }
-
-    pub(crate) fn scan_line(&mut self, line: &str) {
-        self.escaped = false;
-        for (index, ch) in line.char_indices() {
-            match self.quote {
-                Some(RawQuoteKind::Single) => {
-                    if ch == '\'' {
-                        self.quote = None;
-                    }
-                }
-                Some(RawQuoteKind::Double) => {
-                    if ch == '"' && !self.escaped {
-                        self.quote = None;
-                    }
-                }
-                _ => {
-                    if ch == '#' && shell_comment_can_start(line, index) {
-                        break;
-                    }
-                    if ch == '\'' || (ch == '"' && !self.escaped) {
-                        self.quote = Some(if ch == '\'' {
-                            RawQuoteKind::Single
-                        } else {
-                            RawQuoteKind::Double
-                        });
-                    }
-                }
-            }
-
-            self.escaped = ch == '\\' && !self.escaped;
-            if ch != '\\' {
-                self.escaped = false;
-            }
-        }
-        self.escaped = false;
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(crate) struct RawShellScanner<'source> {
-    source: &'source str,
-    upper: usize,
-}
-
-impl<'source> RawShellScanner<'source> {
-    pub(crate) fn new(source: &'source str) -> Self {
-        Self::bounded(source, source.len())
-    }
-
-    pub(crate) fn bounded(source: &'source str, upper: usize) -> Self {
-        Self {
-            source,
-            upper: upper.min(source.len()),
-        }
-    }
-
-    pub(crate) fn skip_single_quoted_body(&self, mut offset: usize) -> usize {
-        while offset < self.upper {
-            let Some(ch) = self.source[offset..].chars().next() else {
-                break;
-            };
-            offset += ch.len_utf8();
-            if ch == '\'' {
-                break;
-            }
-        }
-        offset
-    }
-
-    pub(crate) fn skip_double_quoted_body(&self, mut offset: usize) -> usize {
-        while offset < self.upper {
-            let Some(ch) = self.source[offset..].chars().next() else {
-                break;
-            };
-            offset += ch.len_utf8();
-            if ch == '\\' {
-                if let Some(escaped) = self.source[offset..].chars().next() {
-                    offset += escaped.len_utf8();
-                }
-            } else if ch == '"' {
-                break;
-            }
-        }
-        offset
-    }
-
-    pub(crate) fn skip_escaped_or_quoted_at(&self, offset: usize) -> Option<usize> {
-        let ch = self.source[offset..].chars().next()?;
-        let next = offset + ch.len_utf8();
-        match ch {
-            '\\' => Some(
-                self.source[next..self.upper]
-                    .chars()
-                    .next()
-                    .map_or(next, |escaped| next + escaped.len_utf8()),
-            ),
-            '\'' => Some(self.skip_single_quoted_body(next)),
-            '"' => Some(self.skip_double_quoted_body(next)),
-            _ => None,
-        }
-    }
-
-    pub(crate) fn skip_quoted_or_comment_at(&self, offset: usize) -> Option<usize> {
-        self.skip_escaped_or_quoted_at(offset).or_else(|| {
-            self.comment_starts_at(offset)
-                .then(|| self.comment_end_from(offset))
-        })
-    }
-
-    pub(crate) fn find_comment(&self, start: usize, upper: usize) -> Option<usize> {
-        let mut offset = start.min(self.upper);
-        let upper = upper.min(self.upper);
-        while offset < upper {
-            if let Some(next) = self.skip_escaped_or_quoted_at(offset) {
-                offset = next;
-                continue;
-            }
-
-            let ch = self.source[offset..].chars().next()?;
-            if ch == '#' && shell_comment_can_start(self.source, offset) {
-                return Some(offset);
-            }
-            offset += ch.len_utf8();
-        }
-        None
-    }
-
-    pub(crate) fn is_escaped(&self, index: usize) -> bool {
-        let mut backslashes = 0usize;
-        for ch in self.source[..index.min(self.source.len())].chars().rev() {
-            if ch == '\\' {
-                backslashes += 1;
-            } else {
-                break;
-            }
-        }
-        backslashes % 2 == 1
-    }
-
-    pub(crate) fn matching_command_substitution_close(&self, body_start: usize) -> Option<usize> {
-        let mut quote = QuoteState::default();
-        let mut paren_depth = 0usize;
-        let mut index = body_start.min(self.upper);
-
-        while index < self.upper {
-            let ch = self.source[index..].chars().next()?;
-            let next_index = index + ch.len_utf8();
-            if quote.consume_raw_char(ch, false) {
-                index = next_index;
-                continue;
-            }
-
-            match ch {
-                '(' => paren_depth += 1,
-                ')' => {
-                    if paren_depth == 0 {
-                        return Some(index);
-                    }
-                    paren_depth -= 1;
-                }
-                _ => {}
-            }
-
-            index = next_index;
-        }
-
-        None
-    }
-
-    pub(crate) fn next_command_substitution(&self, mut index: usize) -> Option<(usize, usize)> {
-        let bytes = self.source.as_bytes();
-        let mut quote = QuoteState::default();
-
-        while index + 1 < self.upper {
-            let ch = self.source[index..].chars().next()?;
-            let next_index = index + ch.len_utf8();
-            if quote.consume_shell_word_char(ch) {
-                index = next_index;
-                continue;
-            }
-
-            if !quote.in_single_quotes()
-                && bytes[index] == b'$'
-                && bytes[index + 1] == b'('
-                && bytes.get(index + 2).is_none_or(|byte| *byte != b'(')
-                && let Some(close_offset) = self.matching_command_substitution_close(index + 2)
-            {
-                return Some((index, close_offset));
-            }
-            index = next_index;
-        }
-
-        None
-    }
-
-    pub(crate) fn has_unclosed_substitution_before(&self, upper: usize) -> bool {
-        let mut depth = 0usize;
-        let mut index = 0usize;
-        let upper = upper.min(self.upper);
-        let mut in_single_quotes = false;
-        let mut in_double_quotes = false;
-
-        while index < upper {
-            let Some(ch) = self.source[index..].chars().next() else {
-                break;
-            };
-            let next_index = index + ch.len_utf8();
-
-            if ch == '\\' {
-                index = self.source[next_index..upper]
-                    .chars()
-                    .next()
-                    .map_or(next_index, |escaped| next_index + escaped.len_utf8());
-                continue;
-            }
-            if ch == '\'' && !in_double_quotes {
-                in_single_quotes = !in_single_quotes;
-                index = next_index;
-                continue;
-            }
-            if ch == '"' && !in_single_quotes {
-                in_double_quotes = !in_double_quotes;
-                index = next_index;
-                continue;
-            }
-            if in_single_quotes {
-                index = next_index;
-                continue;
-            }
-
-            let bytes = self.source.as_bytes();
-            match ch {
-                '$' | '<' | '>' if bytes.get(index + 1) == Some(&b'(') => {
-                    depth += 1;
-                    index = index.saturating_add(2);
-                    continue;
-                }
-                ')' if depth > 0 => depth -= 1,
-                _ => {}
-            }
-            index = next_index;
-        }
-
-        depth > 0
-    }
-
-    fn comment_starts_at(&self, offset: usize) -> bool {
-        self.source[offset..self.upper].starts_with('#')
-            && shell_comment_can_start(self.source, offset)
-    }
-
-    fn comment_end_from(&self, offset: usize) -> usize {
-        self.source[offset..self.upper]
-            .find('\n')
-            .map_or(self.upper, |newline| offset + newline + 1)
-    }
-}
-
 pub(crate) fn skip_single_quoted(source: &str, offset: usize, upper: usize) -> usize {
     RawShellScanner::bounded(source, upper).skip_single_quoted_body(offset)
 }
@@ -749,9 +371,10 @@ mod tests {
 
     #[test]
     fn raw_scanner_finds_shell_comments_outside_quotes() {
-        let scanner = RawShellScanner::new("echo '# no' \"# no\" value#no # yes");
+        let source = "echo '# no' \"# no\" value#no # yes";
+        let scanner = RawShellScanner::new(source);
 
-        assert_eq!(scanner.find_comment(0, scanner.source.len()), Some(28));
+        assert_eq!(scanner.find_comment(0, source.len()), Some(28));
     }
 
     #[test]
