+++ crates/shuck-linter/src/facts/word_spans/backticks.rs
@@ -405,174 +405,7 @@ pub(crate) fn skip_shell_whitespace(bytes: &[u8], index: &mut usize) {
 }
 
 pub(crate) fn shell_word_end(text: &str, start: usize) -> Option<usize> {
-    let bytes = text.as_bytes();
-    let mut index = start;
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-
-    while index < bytes.len() {
-        let byte = bytes[index];
-        if in_single_quote {
-            if byte == b'\'' {
-                in_single_quote = false;
-            }
-            index += 1;
-            continue;
-        }
-
-        if byte == b'\\' {
-            index = advance_escaped_shell_char(text, index);
-            continue;
-        }
-
-        if !in_double_quote && byte.is_ascii_whitespace() {
-            break;
-        }
-
-        match byte {
-            b'\'' if !in_double_quote => {
-                in_single_quote = true;
-                index += 1;
-            }
-            b'"' => {
-                in_double_quote = !in_double_quote;
-                index += 1;
-            }
-            b'$' if bytes.get(index + 1) == Some(&b'(') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'(', b')')?;
-            }
-            b'$' if bytes.get(index + 1) == Some(&b'{') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'{', b'}')?;
-            }
-            b'<' | b'>' if !in_double_quote && bytes.get(index + 1) == Some(&b'(') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'(', b')')?;
-            }
-            b'`' => {
-                index = skip_legacy_backtick_construct(text, index + 1)?;
-            }
-            _ => index = advance_shell_char(text, index),
-        }
-    }
-
-    (!in_single_quote && !in_double_quote).then_some(index)
-}
-
-pub(crate) fn skip_balanced_shell_construct(
-    text: &str,
-    mut index: usize,
-    open: u8,
-    close: u8,
-) -> Option<usize> {
-    let bytes = text.as_bytes();
-    let mut depth = 1usize;
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-
-    while index < bytes.len() {
-        let byte = bytes[index];
-        if in_single_quote {
-            if byte == b'\'' {
-                in_single_quote = false;
-            }
-            index += 1;
-            continue;
-        }
-
-        if byte == b'\\' {
-            index = advance_escaped_shell_char(text, index);
-            continue;
-        }
-
-        match byte {
-            b'\'' if !in_double_quote => {
-                in_single_quote = true;
-                index += 1;
-            }
-            b'"' => {
-                in_double_quote = !in_double_quote;
-                index += 1;
-            }
-            b'$' if bytes.get(index + 1) == Some(&b'(') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'(', b')')?;
-            }
-            b'$' if bytes.get(index + 1) == Some(&b'{') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'{', b'}')?;
-            }
-            b'<' | b'>' if !in_double_quote && bytes.get(index + 1) == Some(&b'(') => {
-                index = skip_balanced_shell_construct(text, index + 2, b'(', b')')?;
-            }
-            b'`' => {
-                index = skip_legacy_backtick_construct(text, index + 1)?;
-            }
-            _ if byte == open && !in_double_quote => {
-                depth += 1;
-                index += 1;
-            }
-            _ if byte == close && !in_double_quote => {
-                depth -= 1;
-                index += 1;
-                if depth == 0 {
-                    return Some(index);
-                }
-            }
-            _ => index = advance_shell_char(text, index),
-        }
-    }
-
-    None
-}
-
-pub(crate) fn skip_legacy_backtick_construct(text: &str, mut index: usize) -> Option<usize> {
-    let bytes = text.as_bytes();
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-
-    while index < bytes.len() {
-        let byte = bytes[index];
-        if in_single_quote {
-            if byte == b'\'' {
-                in_single_quote = false;
-            }
-            index += 1;
-            continue;
-        }
-
-        if byte == b'\\' {
-            index = advance_escaped_shell_char(text, index);
-            continue;
-        }
-
-        match byte {
-            b'\'' if !in_double_quote => {
-                in_single_quote = true;
-                index += 1;
-            }
-            b'"' => {
-                in_double_quote = !in_double_quote;
-                index += 1;
-            }
-            b'`' if !in_double_quote => return Some(index + 1),
-            _ => index = advance_shell_char(text, index),
-        }
-    }
-
-    None
-}
-
-pub(crate) fn advance_escaped_shell_char(text: &str, index: usize) -> usize {
-    let next = advance_shell_char(text, index);
-    if next < text.len() {
-        advance_shell_char(text, next)
-    } else {
-        next
-    }
-}
-
-pub(crate) fn advance_shell_char(text: &str, index: usize) -> usize {
-    text[index..]
-        .chars()
-        .next()
-        .map_or(index + 1, |ch| index + ch.len_utf8())
+    shuck_ast::raw_shell::RawShellScanner::new(text).shell_word_end(start)
 }
 
 pub(crate) fn simple_assignment_word(word: &str) -> bool {
